requirements.txt
paddlepaddle>=2.2
skimage
pandas
numpy
matplotlib
sklearn
tqdm